+++ pubsub/adapters/base.py
@@ -9,7 +9,13 @@ class BaseAdapter(object):
     def bulk_publish(self, channel, messages):
         raise NotImplementedError('Not implemented')
 
+    def get_topics(self):
+        raise NotImplementedError('Not implemented')
+
     def subscribe(self, channel, callback, create_topic=False):
+        raise NotImplementedError('Not implemented')
+
+    def delete_topic(self, topic):
         raise NotImplementedError('Not implemented')
 
     def ack(self, message):