+++ pubsub/adapters/googlecloud.py
@@ -1,6 +1,7 @@
 import functools
 import logging
 
+from requests import Session
 from google.api_core.exceptions import GoogleAPICallError
 from google.cloud import pubsub_v1
 from grpc import StatusCode
@@ -16,7 +17,10 @@ class GooglePubsub(BaseAdapter):
     Google-cloud adapter class
     """
 
-    def __init__(self, project_id, client_identifier='default'):
+    def __init__(self, project_id, client_identifier='default', pubsub_rest_proxy=None):
+        if pubsub_rest_proxy:
+            self.pubsub_rest_proxy = pubsub_rest_proxy
+            self.session = Session()
         self.publisher = pubsub_v1.PublisherClient()
         self.subscriber = pubsub_v1.SubscriberClient()
         self.client_identifier = client_identifier
@@ -26,15 +30,31 @@ class GooglePubsub(BaseAdapter):
         message.ack()
 
     def publish(self, topic_name, message):
-        topic_path = self.publisher.topic_path(self.project_id, topic_name)
-        self.get_topic(topic_path)
-        self.publisher.publish(topic_path, message)
+        if self.pubsub_rest_proxy:
+            self.bulk_publish(topic_name, [message])
+        else:
+            topic_path = self.publisher.topic_path(self.project_id, topic_name)
+            self.get_topic(topic_path)
+            self.publisher.publish(topic_path, message)
+
+    def bulk_publish(self, topic_name, messages):
+        if self.pubsub_rest_proxy:
+            # We could use json=, but the message is already dumped and encoded
+            response = self.session.post(
+                '{}/messages/{}'.format(self.pubsub_rest_proxy, topic_name),
+                data=messages, headers={'Content-Type': 'application/json'})
+            response.raise_for_status()
+            return response.json()
+        else:
+            for message in messages:
+                self.publish(topic_name, message)
 
     def subscribe(self, topic_name, callback, create_topic=False):
         # This makes sure the subscription exists
         self.get_subscription(topic_name, create_topic)
 
-        subscription_path = self.subscriber.subscription_path(self.project_id, '{}.{}'.format(self.client_identifier, topic_name))
+        subscription_path = self.subscriber.subscription_path(
+            self.project_id, '{}.{}'.format(self.client_identifier, topic_name))
 
         log.info('Starting to listen on: %s', subscription_path)
         policy = self.subscriber.subscribe(subscription_path)
@@ -44,7 +64,8 @@ class GooglePubsub(BaseAdapter):
         if not self.client_identifier:
             raise IdentifierRequiredException("Use obj.set_client_identifier('name')")
 
-        subscription_path = self.subscriber.subscription_path(self.project_id, '{}.{}'.format(self.client_identifier, topic_name))
+        subscription_path = self.subscriber.subscription_path(self.project_id, '{}.{}'.format(
+            self.client_identifier, topic_name))
         try:
             return self.subscriber.get_subscription(subscription_path)
         except GoogleAPICallError as exc:
@@ -71,7 +92,8 @@ class GooglePubsub(BaseAdapter):
             raise TopicNotFound("Can't delete unknown topic: {}".format(topic_path))
 
     def delete_subscription(self, topic_name):
-        subscription_path = self.subscriber.subscription_path(self.project_id, '{}.{}'.format(self.client_identifier, topic_name))
+        subscription_path = self.subscriber.subscription_path(
+            self.project_id, '{}.{}'.format(self.client_identifier, topic_name))
         try:
             self.subscriber.delete_subscription(subscription_path)
         except GoogleAPICallError as exc: